Putting test classes back into a single file
