--- src/test/java/com/yahoo/bullet/kafka/InstantiableCertRefreshingSSLEngineFactory.java
@@ -1,15 +1,0 @@
-/*
- *  Copyright 2021, Yahoo Inc.
- *  Licensed under the terms of the Apache License, Version 2.0.
- *  See the LICENSE file associated with the project for terms.
- */
-package com.yahoo.bullet.kafka;
-
-import javax.net.ssl.SSLContext;
-
-public class InstantiableCertRefreshingSSLEngineFactory extends CertRefreshingSSLEngineFactory {
-    @Override
-    protected SSLContext createSSLContext() {
-        return null;
-    }
-}
